improve viper abstraction and provide new with option
