+++ service/pkg/config/config.go
@@ -2,19 +2,12 @@ package config
 
 import (
 	"errors"
-	"fmt"
 	"log/slog"
-	"os"
-	"strings"
 
-	"github.com/creasty/defaults"
-	"github.com/fsnotify/fsnotify"
-	"github.com/go-playground/validator/v10"
 	"github.com/opentdf/platform/service/internal/server"
 	"github.com/opentdf/platform/service/logger"
 	"github.com/opentdf/platform/service/pkg/db"
 	"github.com/opentdf/platform/service/tracing"
-	"github.com/spf13/viper"
 )
 
 type ConfigServices map[string]ServiceConfig
@@ -110,117 +103,12 @@ func (c SDKConfig) LogValue() slog.Value {
 	)
 }
 
-// ConfigLoader defines the interface for loading and managing configuration
-type ConfigLoader interface {
-	// Load loads the configuration into the provided struct
-	Load(cfg *Config) error
-
-	// Watch starts watching for configuration changes
-	Watch(cfg *Config) (func(func(fsnotify.Event)), error)
-}
-
-// ViperLoader implements ConfigLoader using Viper
-type ViperLoader struct {
-	viper *viper.Viper
-}
-
-// NewViperLoader creates a new Viper-based configuration loader
-func NewViperLoader(key, file string) (*ViperLoader, error) {
-	homedir, err := os.UserHomeDir()
-	if err != nil {
-		return nil, errors.Join(err, ErrLoadingConfig)
-	}
-
-	// Set paths and config file info
-	v := viper.NewWithOptions(viper.WithLogger(slog.Default()))
-	v.AddConfigPath(fmt.Sprintf("%s/."+key, homedir))
-	v.AddConfigPath("." + key)
-	v.AddConfigPath(".")
-	v.SetConfigName(key)
-	v.SetConfigType("yaml")
-
-	// Default config values (non-zero)
-	v.SetDefault("server.auth.cache_refresh_interval", "15m")
-
-	// Environment variable settings
-	v.SetEnvPrefix(key)
-	v.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
-	v.AutomaticEnv()
-
-	// Allow for a custom config file to be passed in
-	// This takes precedence over the AddConfigPath/SetConfigName
-	if file != "" {
-		v.SetConfigFile(file)
-	}
-
-	// Read the config file
-	if err := v.ReadInConfig(); err != nil {
-		return nil, errors.Join(err, ErrLoadingConfig)
-	}
-
-	return &ViperLoader{viper: v}, nil
-}
-
-// Load loads the configuration into the provided struct
-func (l *ViperLoader) Load(cfg *Config) error {
-	// Set defaults
-	if err := defaults.Set(cfg); err != nil {
-		return errors.Join(err, ErrSettingConfig)
-	}
-
-	// Unmarshal config
-	if err := l.viper.Unmarshal(cfg); err != nil {
-		return errors.Join(err, ErrUnmarshallingConfig)
-	}
-
-	// Validate config
-	validate := validator.New()
-	if err := validate.Struct(cfg); err != nil {
-		return errors.Join(err, ErrUnmarshallingConfig)
-	}
-
-	return nil
-}
-
-// Watch starts watching for configuration changes
-func (l *ViperLoader) Watch(cfg *Config) (func(func(ConfigServices)), error) {
-	l.viper.WatchConfig()
-
-	// Create a slice to store all the hook functions
-	var configChangeHooks []func(ConfigServices)
-
-	// Return a function that allows registering hooks
-	onConfigChange := func(hook func(ConfigServices)) {
-		configChangeHooks = append(configChangeHooks, hook)
-	}
-
-	// Register only one viper config change handler
-	l.viper.OnConfigChange(func(e fsnotify.Event) {
-		slog.Info("Config file changed", "file", e.Name)
-
-		// First reload and validate the config
-		if err := l.Load(cfg); err != nil {
-			slog.Error("Error reloading config", "error", err)
-			return
-		}
-
-		slog.Info("Config successfully reloaded", "config", cfg.LogValue())
-
-		// Then execute all registered hooks with the event
-		for _, hook := range configChangeHooks {
-			hook(cfg.Services)
-		}
-	})
-
-	return onConfigChange, nil
-}
-
 // LoadConfig loads configuration using the provided loader or creates a default Viper loader
 func LoadConfig(key, file string) (*Config, func(func(ConfigServices)), error) {
 	config := &Config{}
 
 	// Create default loader if none provided
-	loader, err := NewViperLoader(key, file)
+	loader, err := NewEnvironmentLoader(key, file)
 	if err != nil {
 		return nil, nil, err
 	}
