+++ service/pkg/server/options.go
@@ -2,6 +2,7 @@ package server
 
 import (
 	"github.com/casbin/casbin/v2/persist"
+	"github.com/opentdf/platform/service/pkg/config"
 	"github.com/opentdf/platform/service/pkg/serviceregistry"
 )
 
@@ -16,6 +17,7 @@ type StartConfig struct {
 	extraCoreServices     []serviceregistry.IService
 	extraServices         []serviceregistry.IService
 	casbinAdapter         persist.Adapter
+	configLoaders         []config.ConfigLoader
 }
 
 // Deprecated: Use WithConfigKey
@@ -98,6 +100,14 @@ func WithServices(services ...serviceregistry.IService) StartOptions {
 func WithCasbinAdapter(adapter persist.Adapter) StartOptions {
 	return func(c StartConfig) StartConfig {
 		c.casbinAdapter = adapter
+		return c
+	}
+}
+
+// WithAdditionalConfigLoader option adds an additional configuration loader to the server.
+func WithAdditionalConfigLoader(loader config.ConfigLoader) StartOptions {
+	return func(c StartConfig) StartConfig {
+		c.configLoaders = append(c.configLoaders, loader)
 		return c
 	}
 }
